+++ src/main/java/step3/RacingController.java
@@ -1,4 +1,4 @@
-package racing;
+package step3;
 
 public class RacingController {
