+++ src/main/java/step3/ResultView.java
@@ -1,4 +1,4 @@
-package racing;
+package step3;
 
 import java.util.List;
 import java.util.Map;
