+++ src/test/java/step3/RacingCarTest.java
@@ -1,4 +1,4 @@
-package racing;
+package step3;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
